[MIZUKI & ADRIAN] Added ResponseEntity in CompanyController.java

+++ src/main/java/com/tw/apistackbase/controller/CompanyController.java
@@ -3,6 +3,8 @@ package com.tw.apistackbase.controller;
 import com.tw.apistackbase.core.Company;
 import com.tw.apistackbase.repository.CompanyRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Optional;
@@ -18,29 +20,39 @@ public class CompanyController {
         return companyRepository.findAll();
     }
 
+    @GetMapping(path = "/{name}", produces = {"application/json"})
+    public Company listById(@PathVariable String name) {
+        return companyRepository.findOneByName(name);
+    }
+
     @PostMapping(produces = {"application/json"})
+    @ResponseStatus(code = HttpStatus.CREATED)
     public Company add(@RequestBody Company company) {
         return companyRepository.save(company);
     }
 
     @PutMapping(path = "/edit/{id}", produces = {"application/json"})
-    public Company edit(@PathVariable Long id, @RequestBody Company company) {
-        Company companyFromDb = companyRepository.findById(id).orElse(null);
+    public ResponseEntity<Company> edit(@PathVariable Long id, @RequestBody Company company) {
+        Company companyFromDB = companyRepository.findById(id).orElse(null);
 
-        if(companyFromDb != null){
-            companyFromDb.setProfile(company.getProfile());
-            companyFromDb.setEmployees(company.getEmployees());
-            companyFromDb.setName(company.getName());
-
-            return companyRepository.save(companyFromDb);
+        if(companyFromDB != null){
+            companyFromDB.setName(company.getName());
+            companyRepository.save(companyFromDB);
+            return new ResponseEntity<>(companyFromDB, HttpStatus.OK);
         }
-        return null;
+
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @DeleteMapping(path = "/delete/{id}", produces = {"application/json"})
-    public Iterable<Company> delete(@PathVariable Long id) {
-        companyRepository.deleteById(id);
-        return companyRepository.findAll();
-    }
+    public ResponseEntity<Company> delete(@PathVariable Long id) {
+        Company companyFromDB = companyRepository.findById(id).orElse(null);
 
+        if(companyFromDB != null) {
+            companyRepository.deleteById(id);
+            return new ResponseEntity<>(companyFromDB, HttpStatus.OK);
+        }
+
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
 }
